src/requirements.txt
requests
dnspython